carrito de compras, facturacion, sql, permisos rol UI

+++ src/main/java/com/clinica/dao/ProductoDao.java
@@ -3,5 +3,5 @@ package com.clinica.dao;
 import com.clinica.domain.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoDao extends JpaRepository<Producto, Integer> {
+public interface ProductoDao extends JpaRepository<Producto, Long> {
 }
